+++ Runtime/Core/Processors/SherpaOnnxUnity/SherpaKeywordDetector.cs
@@ -1,10 +1,12 @@
 #if EASYMIC_SHERPA_ONNX_INTEGRATION
 
-namespace Eitan.EasyMic.Runtime.SherpaOnnxUnity
+namespace Eitan.EasyMic.Runtime.SherpaONNXUnity
 {
     using System;
     using System.Threading;
-    using Eitan.SherpaOnnxUnity.Runtime;
+    using Eitan.SherpaONNXUnity.Runtime.Modules;
+
+
     public class SherpaKeywordDetector : AudioReader, IDisposable
     {