+++ Samples~/Playback Example/Scripts/EasyMicPlaybackExample.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Collections;
-using Eitan.EasyMic.Runtime;
+using Eitan.EasyMic.Runtime.Mono;
 using UnityEngine;
 using UnityEngine.UI;